+++ src/test/java/picocli/MapOptionsTest.java
@@ -5,6 +5,7 @@ import org.junit.Test;
 import org.junit.contrib.java.lang.system.ProvideSystemProperty;
 import org.junit.contrib.java.lang.system.RestoreSystemProperties;
 import org.junit.rules.TestRule;
+import picocli.CommandLine.Model.ArgSpec;
 import picocli.CommandLine.Option;
 import picocli.CommandLine.ParameterException;
 
@@ -29,9 +30,35 @@ public class MapOptionsTest {
     public final ProvideSystemProperty ansiOFF = new ProvideSystemProperty("picocli.ansi", "false");
 
     @Test
-    public void testEmptyStringIfNoValue() {
+    public void testErrorIfNoMapFallbackValue() {
         class App {
             @Option(names = "-D") Map<String, String> map;
+        }
+        try {
+            CommandLine.populateCommand(new App(), "-Dkey");
+            fail("Expected exception");
+        } catch (ParameterException ex) {
+            assertEquals("Value for option option '-D' (<String=String>) should be in KEY=VALUE format but was key", ex.getMessage());
+        }
+    }
+
+    @Test
+    public void testErrorIfMapFallbackValueIsUnspecified() {
+        class App {
+            @Option(names = "-D", mapFallbackValue = "__unspecified__") Map<String, String> map;
+        }
+        try {
+            CommandLine.populateCommand(new App(), "-Dkey");
+            fail("Expected exception");
+        } catch (ParameterException ex) {
+            assertEquals("Value for option option '-D' (<String=String>) should be in KEY=VALUE format but was key", ex.getMessage());
+        }
+    }
+
+    @Test
+    public void testMapFallbackValueEmptyString() {
+        class App {
+            @Option(names = "-D", mapFallbackValue = "") Map<String, String> map;
         }
         App app = CommandLine.populateCommand(new App(), "-Dkey");
         assertEquals(1, app.map.size());
@@ -39,9 +66,19 @@ public class MapOptionsTest {
     }
 
     @Test
-    public void testEmptyStringIfNoValueMultiple() {
+    public void testMapFallbackValueNull() {
         class App {
-            @Option(names = "-D") Map<String, String> map;
+            @Option(names = "-D", mapFallbackValue = ArgSpec.NULL_VALUE) Map<String, String> map;
+        }
+        App app = CommandLine.populateCommand(new App(), "-Dkey");
+        assertEquals(1, app.map.size());
+        assertEquals(null, app.map.get("key"));
+    }
+
+    @Test
+    public void testMapFallbackValueEmptyStringMultiple() {
+        class App {
+            @Option(names = "-D", mapFallbackValue = "") Map<String, String> map;
         }
         App app = CommandLine.populateCommand(new App(), "-Dkey1", "-Dkey2", "-Dkey3");
         assertEquals(3, app.map.size());
@@ -51,9 +88,21 @@ public class MapOptionsTest {
     }
 
     @Test
-    public void testTypeConversionErrorIfNoValue() {
+    public void testMapFallbackValueNullMultiple() {
         class App {
-            @Option(names = "-D") Map<String, Integer> map;
+            @Option(names = "-D", mapFallbackValue = ArgSpec.NULL_VALUE) Map<String, String> map;
+        }
+        App app = CommandLine.populateCommand(new App(), "-Dkey1", "-Dkey2", "-Dkey3");
+        assertEquals(3, app.map.size());
+        assertEquals(null, app.map.get("key1"));
+        assertEquals(null, app.map.get("key2"));
+        assertEquals(null, app.map.get("key3"));
+    }
+
+    @Test
+    public void testTypeConversionErrorIfValueCannotBeConverted() {
+        class App {
+            @Option(names = "-D", mapFallbackValue = "") Map<String, Integer> map;
         }
         try {
             CommandLine.populateCommand(new App(), "-Dkey");