+++ RELEASE-NOTES.md
@@ -23,9 +23,10 @@ Picocli follows [semantic versioning](http://semver.org/).
 ## <a name="4.6.0-new"></a> New and Noteworthy
 
 ## <a name="4.6.0-fixes"></a> Fixed issues
+* [#1241] API: Add `mapFallbackValue` attribute to `@Options` and `@Parameters` annotations, and corresponding `ArgSpec.mapFallbackValue()`.
 * [#1184] API: Added public methods `Help.Layout::colorScheme`, `Help.Layout::textTable`, `Help.Layout::optionRenderer`, `Help.Layout::parameterRenderer`, and `Help::calcLongOptionColumnWidth`.
 * [#1108] Enhancement: Support `Optional<T>` type for options and positional parameters. Thanks to [Max Rydahl Andersen](https://github.com/maxandersen) for raising this.
-* [#1214] Enhancement: Support Map options with key-only (support `-Dkey` as well as `-Dkey=value`). Thanks to [Max Rydahl Andersen](https://github.com/maxandersen) for raising this.
+* [#1214] Enhancement: Support Map options with key-only (support `-Dkey` as well as `-Dkey=value`). Thanks to [Max Rydahl Andersen](https://github.com/maxandersen) and [David Walluck](https://github.com/dwalluck) for raising this and subsequent discussion.
 * [#1236] Enhancement/bugfix: Fix compiler warnings about `Annotation::getClass` and assignment in `if` condition. Thanks to [nveeser-google](https://github.com/nveeser-google) for the pull request.
 * [#1229] Bugfix: Fix compilation error introduced with fc5ef6de6 (#1184). Thanks to [Andreas Deininger](https://github.com/deining) for the pull request.
 * [#1225] Bugfix: Error message for unmatched positional argument reports incorrect index when value equals a previously matched argument. Thanks to [Vitaly Shukela](https://github.com/vi) for raising this.
@@ -47,7 +48,7 @@ Picocli follows [semantic versioning](http://semver.org/).
 No features were deprecated in this release.
 
 ## <a name="4.6.0-breaking-changes"></a> Potential breaking changes
-This release has no breaking changes.
+Added method `isOptional()` to the `picocli.CommandLine.Model.ITypeInfo` interface.
 
 
 # <a name="4.5.2"></a> Picocli 4.5.2